@RealFYang 2023-03-11
Thanks for filing the upstream backporting to fix the failing tests. I will do a merge once that's integrated. 
Regarding this performance backporting, I would put off integration for a while after several other nasty bugfixes are backported.